Extract code from Donate component, Create CustomDonateField component
