+++ app/api/products/route.js
@@ -5,7 +5,7 @@ export async function GET() {
   try {
     const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
     const prices = await stripe.prices.list({
-      limit: 20,
+      limit: 40,
     });
 
     return NextResponse.json(prices.data);
