+++ app/components/Donate.js
@@ -5,11 +5,10 @@ import axios from 'axios';
 import ChildrenCauseDonate from './ChildrenCauseDonate';
 import MWCauseDonate from './MWCauseDonate';
 import DPCauseDonate from './DPCauseDonate';
+import CustomDonateField from './CustomDonateField';
 
 const Donate = () => {
   const [prices, setPrices] = useState([]);
-  const [amount, setAmount] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
 
   useEffect(() => {
     fetchPrices();
@@ -19,68 +18,6 @@ const Donate = () => {
     const { data } = await axios.get('/api/products');
     setPrices(data);
   };
-
-  const handleCustom = async (e) => {
-    e.preventDefault();
-
-    // improve: add logic for cents
-    const parsedAmount = parseInt(amount, 10); // Parses the string into an integer
-
-    if (Number.isInteger(parsedAmount)) {
-      // If the parsed amount is an integer, proceed with the logic
-
-      const { data } = await axios.post(
-        '/api/checkout/custom',
-        {
-          amount: amount,
-          name: 'Children',
-          description: 'Donate to children cause'
-        },
-        {
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        }
-      );
-      window.location.assign(data);
-    } else {
-      // If the entered amount is not a valid integer, handle the error or notify the user
-      setErrorMessage('Please enter a valid amount');
-    }
-    
-  }
-
-  const customDonateSection = (
-    <div>
-      <div className='flex flex-row justify-center'>
-        <div className=' flex items-center justify-center'>$</div>
-        <input 
-          onChange={(e) => setAmount(e.target.value)}
-          value={amount}
-          placeholder="Enter custom amount"
-          className='rounded-lg border border-black m-2 pl-4'
-        />
-        <button onClick={handleCustom} 
-                className='flex h-4 p-3 
-                      bg-gradient-to-b 
-                      from-blue-400 to-yellow-400  
-                      text-black text-l
-                      hover:text-white 
-                      hover:scale-125
-                      font-bold
-                      rounded-lg
-                      items-center
-                      justify-center
-                      ease-in-out duration-300
-                      place-self-center
-                      m-2
-                      md:h-4 md:p-4 md:text-lg
-                      lg:h-8 lg:p-4 lg:text-lg'>Custom Amount</button>
-      </div>
-      {errorMessage && <p style={{ color: 'red' }}>{errorMessage}</p>}
-    </div>
-  )
-  
 
   return (
     <section className="h-fit mx-8 my-6 grid grid-cols-1 sm:grid-cols-1 md:grid-cols-1 lg:grid-cols-3 gap-8">
@@ -103,7 +40,7 @@ const Donate = () => {
         <div>
           <hr className="border-t border-gray-400 my-4"></hr>
           {prices && <ChildrenCauseDonate prices={prices} />}
-          {customDonateSection}
+          <CustomDonateField cause={'Children'} />
         </div>
       </div>
